Modifying splash screen.
Adding UW resources (graphics) to the solution.  The easy way to make
this modification would have been to simply switch the jpg but this
method leaves the original file untouched but has a lot of code changes
to support it.  This may cause problems in the future if we have a
collision with future changes to the same line of code.